Refactor router and add pre:*, post:* hooks
First of all, the router needed a refactoring, it was calling every middleware by code repetition, honestly it was bad code and it
had to go. Secondly, I wanted to add two more hooks, pre:* and post:* : I often needed to run a certai middleware on any resource endpoint and
while you can easily use connect's app.use() I wanted a more coherent way of doing it with the framework.