Refactor: use a getter to access the resource's router
Previously, the Resource's router was built in the Resource's controller, this broke
the possibility to register hooks after creating the resource. With this commit we create a class property getter on the router attribute,
to ensure that users of the Resource class will always get the latest version of the router.